+++ src/AccessControl/ImplPython.py
@@ -26,13 +26,12 @@ from Acquisition import aq_parent
 from ExtensionClass import Base
 from zope.interface import implementer
 
-# This is used when a permission maps explicitly to no permission.  We
-# try and get this from cAccessControl first to make sure that if both
-# security implementations exist, we can switch between them later.
-try:
-    from AccessControl.cAccessControl import _what_not_even_god_should_do
-except ImportError:
+PURE_PYTHON = int(os.environ.get('PURE_PYTHON', '0'))
+if PURE_PYTHON:
+    # We need our own to not depend on the C implementation:
     _what_not_even_god_should_do = []
+else:
+    from AccessControl.cAccessControl import _what_not_even_god_should_do
 
 from AccessControl.interfaces import ISecurityManager
 from AccessControl.interfaces import ISecurityPolicy
@@ -42,6 +41,8 @@ from AccessControl.SimpleObjectPolicies import Containers
 from AccessControl.SimpleObjectPolicies import _noroles
 from AccessControl.unauthorized import Unauthorized
 from AccessControl.ZopeGuards import guarded_getitem  # NOQA
+
+
 # AccessControl.ZopeSecurityPolicy
 # --------------------------------
 #
@@ -76,8 +77,8 @@ def rolesForPermissionOn(perm, object, default=_default_roles, n=None):
             roles = getattr(object, n)
             if roles is None:
                 if _embed_permission_in_roles:
-                    return ('Anonymous', n)
-                return 'Anonymous'
+                    return (('Anonymous',), n)
+                return ('Anonymous',)
 
             t = type(roles)
             if t is tuple: